// Frontend/src/redux/types.ts
export enum types{
    GET_ALL_COUNTRIES = "GET_ALL_COUNTRIES",
    GET_ALL_USERS ="GET_ALL_USERS",
    GET_ALL_ROOMS_BY_HOTEL= "GET_ALL_ROOMS_BY_HOTEL",
    GET_ALL_HOTELS = "GET_ALL_HOTELS",
    GET_RESERVATION_BY_USER ="GET_RESERVATION_BY_USER",
    GET_RESERVATION_BY_ID ="GET_RESERVATION_BY_ID",
    POST_RESERVATION = "POST_RESERVATION",
    UPDATE_RESERVATION = "UPDATE_RESERVATION",
    CANCEL_RESERVATION = "CANCEL_RESERVATION",
    RESET = "RESET"
}